Dismantle weird AccountSelector component, fix IOU txn state mgmt
