+++ client/appSrc/components/utils/classifyTransactions.js
@@ -1,12 +1,11 @@
 /* eslint-disable camelcase */
 const classifyTransactions = ({
   displayedTransactions,
-  sharedAccounts,
-  sharedParentCategories,
-  // splitAccountId,
+  selectedAccounts,
+  selectedParentCategories,
 }) => {
-  const sharedAccountIds = sharedAccounts.map((acct) => acct.accountId);
-  const sharedCategoryIds = sharedParentCategories
+  const sharedAccountIds = selectedAccounts.map((acct) => acct.accountId);
+  const sharedCategoryIds = selectedParentCategories
     .flatMap(({ subCategories }) => subCategories)
     .map(({ id }) => id);
 
@@ -21,19 +20,15 @@ const classifyTransactions = ({
     const isInSharedAccountButNotCategory = isInSharedAccount && !isInSharedCategory;
     const isInSharedCategoryButNotAccount = isInSharedCategory && !isInSharedAccount;
 
-    // if (splitAccountId === account_id) accum.iouAccountTransactions.push(transaction);
-    if (isInSharedAccount) accum.transactionsInSharedBankAccounts.push(transaction);
     if (isInSharedCategory) accum.transactionsInSharedCategories.push(transaction);
     if (isInSharedAccountButNotCategory) accum.sharedAccountErrorTransactions.push(transaction);
     if (isInSharedCategoryButNotAccount) accum.sharedCategoryErrorTransactions.push(transaction);
 
     return accum;
   }, {
-    transactionsInSharedBankAccounts: [],
     transactionsInSharedCategories: [],
     sharedAccountErrorTransactions: [],
     sharedCategoryErrorTransactions: [],
-    // iouAccountTransactions: [],
   });
 };
